+++ flagsmith/resource_feature_state_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 )
 
-func TestAccFlagResource(t *testing.T) {
+func TestAccFeatureStateResource(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,